requirements.txt
azure-storage-blob 
streamlit
docx2txt
pypdf
langchain-community
langchain-text-splitters
python-dotenv
langchain-chroma
langchain-openai
unstructured
langchain
python-pptx
boto3
